refactor: router 분리 및 recoil 설치
