+++ front/package.json
@@ -21,6 +21,7 @@
     "react-cookie": "^6.1.1",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.18.0",
+    "recoil": "^0.7.7",
     "styled-components": "^6.1.0",
     "styled-reset": "^4.5.1",
     "three": "^0.158.0",
