+++ front/vite.config.ts
@@ -1,15 +1,9 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react-swc';
 import mkcert from 'vite-plugin-mkcert';
-import { compressNormals, compressPositions } from 'three/examples/jsm/utils/GeometryCompressionUtils.js';
-import { compression } from 'vite-plugin-compression2'
 
 export default defineConfig({
-  plugins: [
-    react(),
-    mkcert(),
-
-  ],
+  plugins: [react(), mkcert()],
   server: {
     host: '0.0.0.0',
     https: true,
@@ -36,11 +30,11 @@ export default defineConfig({
     rollupOptions: {
       output: {
         manualChunks: {
-          three : ['three'],
-          'three-fiber' : ['@react-three/fiber'],
-          'three-drei' : ['@react-three/drei'],
-        },
-      },
-    },
-  },
+          three: ['three'],
+          'three-fiber': ['@react-three/fiber'],
+          'three-drei': ['@react-three/drei']
+        }
+      }
+    }
+  }
 });
